requirements.txt
rpaframework==11.3.0
pandas==1.3.3
selectorlib==0.16.0